Solution to Streaming Service Subscription Management

+++ Streaming Service Subscription Management@@67877d61dc107486ad5c2752/code.c
@@ -4,7 +4,7 @@ int main(){
     struct Subscription{
         char name[100];
         char typr[100];
-        int price;
+        float price;
     };
     int n;
     scanf("%d",&n);
@@ -12,7 +12,7 @@ int main(){
     for(int i=0; i<n ; i++){
         scanf("%s",&arr[i].name);
         scanf("%s",&arr[i].typr);
-        scanf("%d",&arr[i].price);
+        scanf("%f",&arr[i].price);
     }
     float c=0,b=0,t=0;
     int ub=0 ,us=0,up=0;
